LayoutTests/css3/scroll-snap/scroll-snap-discrete-wheel-event-in-mainframe-expected.txt

This test verifies that discrete wheel events trigger scroll snapping in the mainframe.
        To manually run the test, scroll using a physical mouse wheel and verify that a single scroll
        tick animates to the next snap point, and a second scroll tick should not trigger snapping (due
        to the content being too tall)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS pageYOffset became 600
PASS pageYOffset > 600 became true
PASS pageYOffset < 1500 is true
PASS successfullyParsed is true

TEST COMPLETE
